refactor: 테스트 패키지 정리
