+++ src/test/java/persistence/sql/ddl/CreateQueryBuilderTest.java
@@ -3,7 +3,7 @@ package persistence.sql.ddl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import persistence.example.Person;
-import persistence.sql.NotEntity;
+import persistence.fixture.NotEntity;
 import persistence.sql.meta.Table;
 
 import static org.assertj.core.api.Assertions.*;
