+++ src/test/java/persistence/sql/meta/FieldTypeTest.java
@@ -1,8 +1,7 @@
-package persistence.sql;
+package persistence.sql.meta;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import persistence.sql.meta.FieldType;
 
 import java.lang.reflect.Field;
